+++ src/export/RenderModelGeometry.cpp
@@ -65,15 +65,15 @@ RenderModelGeometry makeBoxRenderModelGeometry(const float edgeMeters)
     geometry.vertices.reserve(24);
     geometry.indices.reserve(36);
     for (std::size_t faceIndex = 0; faceIndex < faces.size(); ++faceIndex) {
-        const std::uint16_t base = static_cast<std::uint16_t>(geometry.vertices.size());
+        const std::uint32_t base = static_cast<std::uint32_t>(geometry.vertices.size());
         for (const int cornerIndex : faces[faceIndex]) {
             RenderModelVertex vertex;
             vertex.position = corners[static_cast<std::size_t>(cornerIndex)];
             vertex.normal = normals[faceIndex];
             geometry.vertices.push_back(vertex);
         }
-        geometry.indices.insert(geometry.indices.end(), {base, static_cast<std::uint16_t>(base + 1), static_cast<std::uint16_t>(base + 2)});
-        geometry.indices.insert(geometry.indices.end(), {base, static_cast<std::uint16_t>(base + 2), static_cast<std::uint16_t>(base + 3)});
+        geometry.indices.insert(geometry.indices.end(), {base, base + 1u, base + 2u});
+        geometry.indices.insert(geometry.indices.end(), {base, base + 2u, base + 3u});
     }
     return geometry;
 }
